+++ src/java/model/tableclasses/GenericItem.java
@@ -10,22 +10,22 @@ package model.tableclasses;
  * @author Conor
  */
 public class GenericItem {
-    private int id;
+    private long id;
     private String name;
 
     public GenericItem() {
     }
 
-    public GenericItem(int id, String name) {
+    public GenericItem(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
